Time: 9 ms (56.53%), Space: 9.5 MB (60.5%) - LeetHub

+++ 0224-basic-calculator/0224-basic-calculator.cpp
@@ -1,6 +1,6 @@
 class Solution {
 public:
-    long long calculate(string s) {
+    int calculate(string s) {
         stack <long long> nums, ops;
         long long num = 0;
         long long res = 0;
